pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[project]
name = "oac"
version = "0.3.0"
description = "Open Agent Capsule starter toolkit"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.11,<3",
  "PyYAML>=6.0.2,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.12,<0.13",
  "maturin>=1.5,<2.0",
]

[project.scripts]
oac = "oac.cli:main"

[tool.maturin]
module-name = "oac_core"
python-source = "src"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
